Prevents RCE by whitelisting allowed types
The XML is a attacker-controlled input. In addition to the values being
provided by the user, this also allows new values to be de-serialized,
leading to remote code execution (e.g. [CVE-2013-7285][0]).

Types are now whitelisted by being either (a) explictly whitelisted
or (b) aliased. The underlying assumption is that any type that has
an aliased name for serialization (compared to: it's fully qualified
name) is intended to be deserialized.

The unit tests cover a number of scenario's and contain a POC that pops
Calculator.app.

[0]: https://x-stream.github.io/CVE-2013-7285.html