Manifest entry file names should only use current directory
Only allow normal characters in manifest entry file names. RFC 6486 is
actually not clear what is allowed and simply requires that manifest
entries are published in the same repository as the manifest. Quote
from the RFC: "the name of the file in the repository publication
point (directory)".

We allow the characters from `DERPrintableString#isPrintableString`
except for slash (directory separator). We also allow the
underscore (`_`) since it is used by some certificate authorities
already.