Use base64 mime decoder to be compatible with XStream decoder
There is a test
case (ProcessApnicPdusTest.apnic_pdu_2011_08_15_1_has_errors) that
breaks when using the Java basic base64 decoder (which complies with
RFC 4648 as required by RFC 6492). It works with the Java MIME base64
decoder (specified by RFC 2045).

For now we'll be permissive when parsing and strict when generating
base64 encoded fields.