The req_resource_set_* attributes are not always required
From the https://tools.ietf.org/html/rfc6492#section-3.3.2 about these
attributes:

```
If this attribute was present in the certificate
request, then the attribute MUST be present in this response;
otherwise, it MUST NOT be present.
```

We use `null` as a "not-present" value. Later we may want to switch to
use more explicit `Optional` types.