Replace joda time `DateTime` with Java time `Instant`
Mostly straightforward replacement, except for
`DateTimeUtils.setCurrentMillisFixed` and
`DateTimeUtils.currentTimeMillis`, which requires the injection of a
`java.time.Clock` instance in `ValidationOptions` and various
builders to allow setting the current time for testing purposes.

In a few places this replacement was hard, so we use the actual
current time instead: `Rsync` (for calculating start and stop time),
`TrustAnchorRequest` (creation time), and adding a metric to a
validation run (measured at time).

Enforces UTC time zone since Java's `Instant` does not have a time
zone.

Breaks compatibility with previous API, so version bumped to
2.0-SNAPSHOT.